added a custom token example
